ttsserver/requirements.txt
azure-cognitiveservices-speech==1.17.0
baidu-aip
boto3
bs4
coloredlogs
flask
lxml
numpy
pinyin==0.2.5
pydub
pysptk==0.1.4
python-slugify>=5.0.2
pyyaml
requests
scipy
sox>=1.2.9
urllib3>1.25.4
websocket-client==1.2.1
